requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
tensorflow==2.20.0
numpy==1.26.5
pillow==10.0.1
pandas==2.2.2
opencv-python==4.9.0.74
sqlalchemy==2.0.22
python-multipart==0.0.6
streamlit==1.26.0
requests==2.31.0
scipy==1.13.1